[RSN-19] - Create Controllers (#79)
* test: add unit tests for jwt, handlers and validators

* feat: add placeholders for endpoints

* chore: add missing controllers placeholders

* Restored GetCurrentUser method

* [RSN-60] - Implement required methods in MeController and UserController (#58)

* feat: add jwt and exceptions handlers

Simplify the logic on the controllers' side, custom exceptions and
a library for validation were added. The implemented handlers are
allow to catch those exceptions and return the corresponding
statuses and detailed response.

Additionally, mappers have been created to more easily convert
entities into the corresponding DTOs, fixed enum conversion
when it comes to UserRole and export postgres port on the
development environment.

* test: add unit tests for jwt, handlers and validators

* feat: add placeholders for endpoints

* chore: add missing controllers placeholders

* Created controllers

* Created controllers

* Finished controllers and updated services

* Update UserService tests

* Create and update image methods update

* Image controllers update

* Update controller and service

* Added Address to UserDto

* Controller update and service changes

Changed Image and Participant service (and participant tests), controller now use new service methods

* Methods name changes

---------

Co-authored-by: raczu <Raczuekk@gmail.com>

* [RSN-59] - Created EventsController (#59)

* feat: add jwt and exceptions handlers

Simplify the logic on the controllers' side, custom exceptions and
a library for validation were added. The implemented handlers are
allow to catch those exceptions and return the corresponding
statuses and detailed response.

Additionally, mappers have been created to more easily convert
entities into the corresponding DTOs, fixed enum conversion
when it comes to UserRole and export postgres port on the
development environment.

* test: add unit tests for jwt, handlers and validators

* feat: add placeholders for endpoints

* chore: add missing controllers placeholders

* Created EventsController

* Little Updates

* Updated controllers

* Update adding and updating image in eventController

* Updated eventcontroller logic

* Updated EventsController

* Fixed build issues

* updated eventscontroller

* fixed build issue

addedscope for address service

* Update EventServicesTest.cs

* Update EventsController.cs

* syntax update

* updated models

changed removing tags and params that are not attached to any event, thay will be hold in database

* Update EventServicesTest.cs

* Updated EventController

* updated events controller

added returnign empty list if image for event is not found, filtered get all events to not return waiting for aproval events

* updated logic in getallaevenst

it wont return events where status is receted canceled and panding aproval

* fixes and updates

* Fixes and cleanup after rebase and confilts

* Update EventService.cs

* updates

* Updated logic with slag on creation

* Update EventValidatorTests.cs

* Update EventValidatorTests.cs

---------

Co-authored-by: raczu <Raczuekk@gmail.com>

* Controller and service update

Return user's image for comment

---------

Co-authored-by: raczu <Raczuekk@gmail.com>
Co-authored-by: Maciej Koperdowski <maciej.koperdowski@gmail.com>
Co-authored-by: Maciej Koperdowski <85642918+mkoper02@users.noreply.github.com>